Discop/requirements.txt
nltk==3.7
datasets==2.0.0
transformers~=4.19.2
scipy~=1.7.3
pandas~=1.3.5
# torch==1.11.0
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
torchaudio==0.11.0
# pip install torch==1.11.0+cu113 torchvision==0.12.0+cu113 torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu113
numpy~=1.21.6
librosa~=0.8.0
tqdm~=4.62.3
omegaconf~=2.0.6
pillow~=9.0.1
cython~=0.29.28
soundfile~=0.10.3.post1
toml==0.10.2
inflect==5.6.0
sacremoses
importlib_resources